+++ src/components/flash/flash.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import shouldComponentUpdate from './../../utils/helpers/should-component-update';
 import I18n from 'i18n-js';
 import classNames from 'classnames';
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import CSSTransitionGroup from 'react-transition-group/CSSTransitionGroup';
 import Icon from './../icon';
 import Alert from './../alert';
 import Link from './../link';
@@ -447,18 +447,18 @@ class Flash extends React.Component {
     return (
       <div { ...tagComponent('flash', this.props) }>
         <div className={ this.classes }>
-          <ReactCSSTransitionGroup
+          <CSSTransitionGroup
             transitionName="carbon-flash__slider"
             transitionEnterTimeout={ 600 }
             transitionLeaveTimeout={ 600 }>
             { sliderHTML }
-            <ReactCSSTransitionGroup
+            <CSSTransitionGroup
               transitionName="carbon-flash__content"
               transitionEnterTimeout={ 800 }
               transitionLeaveTimeout={ 500 } >
               { flashHTML }
-            </ReactCSSTransitionGroup>
-          </ReactCSSTransitionGroup>
+            </CSSTransitionGroup>
+          </CSSTransitionGroup>
         </div>
 
         { this.dialogs }