+++ src/components/modal/modal.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import CSSTransitionGroup from 'react-transition-group/CSSTransitionGroup';
 import Events from './../../utils/helpers/events';
 
 /**
@@ -24,7 +24,7 @@ import Events from './../../utils/helpers/events';
  * get onClosing() // Called by componentDidUpdate when dialog closes
  * get mainClasses() // Classes to apply to parent div
  * get modalHTML() // JSX displayed when open
- * get transitionName() // Transisition name for ReactCSSTransitionGroup
+ * get transitionName() // Transisition name for CSSTransitionGroup
  *
  * Optional Override
  * get backgroundTransitionName() // Transisition name for background fade
@@ -198,19 +198,19 @@ class Modal extends React.Component {
         className={ this.mainClasses }
         { ...this.componentTags(this.props) }
       >
-        <ReactCSSTransitionGroup
+        <CSSTransitionGroup
           transitionName={ this.transitionName }
           transitionEnterTimeout={ 500 }
           transitionLeaveTimeout={ 500 } >
           { modalHTML }
-        </ReactCSSTransitionGroup>
+        </CSSTransitionGroup>
 
-        <ReactCSSTransitionGroup
+        <CSSTransitionGroup
           transitionName={ this.backgroundTransitionName }
           transitionEnterTimeout={ 500 }
           transitionLeaveTimeout={ 500 } >
           { backgroundHTML }
-        </ReactCSSTransitionGroup>
+        </CSSTransitionGroup>
       </div>
     );
   }