+++ src/components/table/table.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import CSSTransitionGroup from 'react-transition-group/CSSTransitionGroup';
 import classNames from 'classnames';
 import Immutable from 'immutable';
 import I18n from 'i18n-js';
@@ -870,7 +870,7 @@ class Table extends React.Component {
     return (
       <TableRow key="__loading__" selectable={ false } highlightable={ false } hideMultiSelect={ true }>
         <TableCell colSpan="42" align="center">
-          <ReactCSSTransitionGroup
+          <CSSTransitionGroup
             transitionName="table-loading"
             transitionEnterTimeout={ 300 }
             transitionLeaveTimeout={ 300 }
@@ -878,7 +878,7 @@ class Table extends React.Component {
             transitionAppear={ true }
           >
             <Spinner size="small" />
-          </ReactCSSTransitionGroup>
+          </CSSTransitionGroup>
         </TableCell>
       </TableRow>
     );